+++ CarverneMerveilleuse/Assets/Script/PlayerCharacter/Weapon/Light/PlayerAttackCollision2.cs
@@ -19,7 +19,7 @@ public class PlayerAttackCollision2 : MonoBehaviour
     private int randAudioHit;
     private int randAudioNoHit;
     
-    [Space] [SerializeField] private int scoreRagePoint = 1;
+    [Space] [SerializeField] private float scoreRagePoint = 1;
     
     public static PlayerAttackCollision2 instance;
